updated results

+++ runtime/Simulations_Figure5.py
@@ -101,6 +101,9 @@ class Simulation:
         dir = self.results_folder[:self.results_folder.rfind('/')]
         Path(dir).mkdir(parents=True,exist_ok=True)
 
+        if(fileExist):
+            return fileExist
+        
         with open(self.results_folder, 'w', encoding='UTF8', newline='') as f:
             writer = csv.writer(f)
             writer.writerow(header)
